configs/frags/armeb: really use the Linaro toolchain
If you don't enable BR2_ARM_ENABLE_VFP=y, then BR2_ARM_EABIHF=y does
not do anything because EABIhf is not available for cores that don't
have a VFP.

Since the Linaro ARM big-endian toolchain is an EABIhf only toolchain,
it is not available if EABI is used instead of EABIhf. Due to this, it
was using the Musl big-endian toolchain, which is big-endian for
ARMv4/5 (i.e BE-32).

To fix this, we define BR2_ARM_ENABLE_VFP=y. In addition, there is a
different Config.in symbol for the ARM little-endian and the ARM
big-endian Linaro toolchain, so we use the appropriate one.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>