util-linux: add util-linux-dmesg to only build dmesg
Add a util-linux-dmesg package to only build dmesg from util-linux.
This is to optimize for size, since there is no granularity in the
standard util-linux package configuration to only select some
binaries.

The UTIL_LINUX_DMESG configuration for this package is mutually
exclusive with UTIL_LINUX_BINARIES to not have it built twice.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>